# Add a generated trie to speed up keyword detection