Performance delta:
```
benchmark                                  old ns/op     new ns/op     delta
BenchmarkKeywordScan/keywords/ATTACH-8     24.0          9.30          -61.25%
BenchmarkKeywordScan/keywords/DROP-8       23.1          10.0          -56.71%
BenchmarkKeywordScan/keywords/SELECT-8     23.9          8.50          -64.44%
BenchmarkKeywordScan/keywords/VACUUM-8     24.5          8.47          -65.43%
```